separate hot reloading example from default one
