+++ examples/default/src/index.js
@@ -1,13 +1,12 @@
 import { app } from "hyperapp";
-import moisturize from "hyperapp-moisturize";
-import state from "./state";
-import actions from "./actions";
-import view from "./view";
-import makeHotReloader from "./makeHotReloader";
+import H1 from "./components/H1";
 
-const main = moisturize(app)(state, actions, view, document.body);
+const state = {
+  message: "hello world"
+};
 
-const hotReloader = makeHotReloader(main.updateApp);
-hotReloader("state.js", "newState");
-hotReloader("actions.js", "newActions");
-hotReloader("view.js", "newView");
+const actions = {};
+
+const view = ({ message }) => H1({}, message);
+
+app(state, actions, view, document.body);
