+++ examples/hot/package.json
@@ -3,6 +3,6 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node ../../src"
+    "start": "node ../../src --hot"
   }
 }
