+++ examples/hot/src/actions.js
