+++ examples/hot/src/components/H1.js
@@ -1,0 +1,3 @@
+import { h } from "hyperapp";
+
+export default h.bind(null, "h1");
