+++ examples/hot/src/makeHotReloader.js
