+++ examples/hot/src/state.js
