+++ examples/hot/src/view.js
